requirements.txt
streamlit>=1.25.0
tensorflow>=2.11.0,<2.14.0
numpy
pandas
scikit-learn
matplotlib
joblib
pillow